+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This is a log of major user-visible changes in each phpMyFAQ release.
 ### phpMyFAQ v3.2.0-dev - 2022-XX-XX
 
 - changed PHP requirement to PHP 8.0+ (Thorsten)
+- migrated from SwiftMailer to Symfony Mailer (Thorsten)
 
 ### phpMyFAQ v3.1.0-RC - 2022-01-22